fix(service-spec): skip adding a resource property that could not be resolved (#1508)
Fixes:
if the CFN schema contains some resource or type that contains a
property of type that could not be resolved (like by mistake the type
definition does not exist in the schema), now the specs importer failed
to continue processing the remaining properties, and types defined in
this resource/type.

This PR is to only skip the property that has issues, and continue
importing everything else.