fix(service-spec): treat evolution of json -> string as a bugfix (#1923)
The type of a property holds a set of `previousTypes`. The reason we
hold this type history is because if a CDK Resource property was ever
typed as "json", we can never change it to a named type, for backwards
compatibility reasons. The L1 code generator for CDK always takes the
oldest type that a property advertises and codegens using that.

(In the future it would be possible to generate multiple fields with
different types, for example, but we're not currently doing that yet.)

There are two reasons why a property type can change:

- Gradual type refinement (some type that used to be typed as `Json` now
gets a named type with accurate field definition)
- Bugfixes (oops! It was never actually an `X` but the schema
erroneously advertised it as such)

The reasons for backwards compatibility exist to deal with the first
situation, but we sometimes have the second situation.

In https://github.com/aws/aws-cdk/issues/26749, which was the motivation
for this change: at some point in the past (in the legacy CFN spec,
specifically) the type of `UpdatedBy` was `Json`, but this got changed
to `String` later. In fact CloudFormation fails deployment if the value
is anything other than a `String`, but our backwards compatibility rules
mandate that because this type was at some point in the past `Json`, it
shall remain `Json` forever.

Because the backwards compatibility rules are intended to deal with
changes of `Json` -> "named type", we add an exception here: if the
change is `Json` -> `String`, we assume this is a bugfix change and we
don't record the old type.

Fixes https://github.com/aws/aws-cdk/issues/26749

Also changed in this PR:

- I noticed some types were marked as a union with 1 branch. In order to
simplify the type detection, I've made it so that these are simplified
to just the 1 type itself.