fix: the primary identifier is not an attribute
The set of attributes of a resource used to be determined as:

```
readOnlyProperties - createOnlyProperties
```

This would reliably filter out primary identifiers like `Name` from
the set of attributes, since those can be configured and lead to
replacement.

However, it doesn't filter out identifiers as the primary identifier:
those are not part of `createOnlyProperties` but should be excluded
equally.

Change the set to

```
readOnlyProperties - createOnlyProperties - [primaryIdentifier]
```