fix: invalid attributes and properties remain because of renames
The `ResourceBuilder` used to mess around moving properties between the
mutable property bags of `resource.properties` and
`resource.attributes`, which would already contain data from previous
versions of the spec.  This could lead to the parsing of a new version
of the spec interfering with the result of a parsing of a previous
version of the spec.

Instead, we now mutate `candidateProperties` and `candidateAttributes`, and
commit the final state to the underlying `Resource` object when we're
done.