fix: invalid attributes and properties remain because of renames (#1996)
The `ResourceBuilder` used to mess around moving properties between the
mutable property bags of `resource.properties` and
`resource.attributes`, which would already contain data from previous
versions of the spec. This could lead to the parsing of a new version of
the spec interfering with the result of a parsing of a previous version
of the spec.

Instead, we now mutate `candidateProperties` and `candidateAttributes`,
and commit the final state to the underlying `Resource` object when
we're done.

* This change surfaced a number of attributes which are still encoded
into the old resource spec as properties. Patch those away.

Strangely enough, this change seems not to surface any other problems in
the spec and names used in the spec. Nevertheless, the new behavior is
easier to reason about, and makes implementing a subsequent PR easier
(https://github.com/cdklabs/awscdk-service-spec/pull/1990), so we're
keeping it.