feat: sanitize primary identifiers (#2011)
There are at least two resource types,
`AWS::ApiGateway::RequestValidator` and `AWS::ApiGateway::Model`, whose
primary identifiers do not match reality. In the schema, the primary
identifier for these resources includes the `RestApiId` attribute. But
this attribute is not returned by the CloudFormation intrinsic `Ref`
function, and it's not possible to `Fn::GettAtt` it either.

Sanitize the `primaryIdentifier` for these resource types, by removing
the `RestApiId` attribute from the array.