feat: update CloudWatch Console Service Directory source (#2028)
Manual one-off update of the source, since we haven't done this for a
while.

This has a lot of changes, but they don't have a huge effect on
`aws-cdk-lib` since only a few L2s currently use them.
And in any case, we consider these part of the automated surface.