feat: add relationshipRefs on resource properties (#2067)
### Changes
Add support for the `relationshipRefs` property for resource properties.
This new type will hold the relationship information between the current
property and multiple pairs of `{ typeName, propertyName }`.

For example if the `AWS::Lambda::Function Role` property has
relationshipRefs `[{ typeName: "AWS::IAM::Role", propertyName: "Arn" }]`
it means that the role property of the lambda function would accept the
arn property of the IAM role

---------

Co-authored-by: Leon Michalski <leonmk@amazon.de>