fix(oob-relationships): skip relationships for description properties (#2124)
Fixes #

### Description
This change remove the relationships on description properties as they
are not useful: description properties are strings that can be set to
the users liking, thus a relationship saying that the description could
be an EC2 subnet id doesn't make sense.

### Changes
Ignore relationships whose property path ends with `/Description` and
add a unit test for it.

---------

Signed-off-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Leon Michalski <leonmk@amazon.de>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>