fix: loading database crashes if collections are missing (#2237)
The database `.load()` function was expecting that the data file would
always conform to the exact (collection) schema that was defined in the
database.

In general this should be true, but there is one exception:

- If we are adding new collections; and
- We are doing a database diff between a new database (with the new
schema) and the currently published version of the database (created
using the old schema)

Instead, we will treat the missing keys as empty.

---------

Co-authored-by: Rico Huijbers <rix0rrr@gmail.com>