fix: do not print event diff when the ids have changed (#2252)
Fixes #2248

### Description of the bug
When checking for the database diff, the `eventRootProperty` and
`eventResourcesField` of events were shown as changing like:
```│       └ events
│          ├[~]  event aws.autoscaling@EC2InstanceLaunchLifecycleAction
│          │  └      - rootProperty: EC2InstanceLaunchLifecycleAction
│          │         + rootProperty: EC2InstanceLaunchLifecycleAction
│          │         - resourcesField: [EC2InstanceLaunchLifecycleAction.AutoScalingGroupName]
│          │         + resourcesField: [EC2InstanceLaunchLifecycleAction.AutoScalingGroupName]
```
This was caused because they used `JsonEq` for comparison, which also
checked if they has the same `$id` in the db, which is not always true
(e.g. when adding a new resource to the db, ids will change).

### Solution
Do not compare the ids, but use the existing functions to compare the
`EvenTypeDefinition` in these properties.

### Testing
- Ran the diff script on databases where the ids had changed, diff is
now correct.
  - Added a unit test

Co-authored-by: Leon Michalski <leonmk@amazon.de>