fix(typewriter): change  type from ThisInstance to Expression (#2254)
For some unknown reason, when operating in a large repo like
[aws/aws-cdk](https://github.com/aws/aws-cdk), using the `$this` helper
causes the type checker to fail and "ignore" the `ExpressionProxy` part
of the type. This then causes a lot of type errors. Only happens in the
IDE and is not consistent.

I tested a few things and changing the type of `$this` from
`ExpressionProxy<ThisInstance>` to `ExpressionProxy<Expression>` seems
to fix it. This is okay, because `ThisInstance` currently just extends
`Expression`. And in any case the helper is intended to be used as a
proxy, not as `ThisInstance` itself.

It's not really a satisfying solution, but a practical one.