fix: override certain incorrect primary identifiers
There are a number of primary identifiers in the Registry Schema that
don't agree with what the `{ Ref }` intrinsic of CloudFormation is
returning.

This is not unexpected or wrong: the Registry Schema describes the
behavior of Cloud Control API, and the identifier uniquely identifies a
resource in an account.

However, certain resources in CloudFormation already had a certain
behavior, and perhaps the `{ Ref }` returned a value that wasn't unique
enough; like the name of a policy inside another resource: unique
*enough* in the context of a CloudFormation template, but not unique
enough for CCAPI.

Since the behavior of CloudFormation cannot be changed, this leads to
the inescapable conclusion that the primary identifiers of CCAPI and
CloudFormation can be different.

CDK needs to account for this difference, so we do, introducing two
kinds of primary identifiers: a CCAPI-level one and a CFN-level one.

We will do most of our codegen work off of the CFN identifier, but
we'll have the other one in our back pocket.

This list is not subject to maintenance (modulo things that Kiro and
myself have missed for whatever reason, I didn't do all of this work by
myself because that would have been crazy) because going forward for new
resources the schema must agree with the behavior of `{ Ref }`.